# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Naiad API
  version: "1.0.0"
  description: Garden irrigation controller — REST API
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SequenceStatus:
      type: string
      enum: [idle, running, paused, disabled]
    TriggerSource:
      type: string
      enum: [cron, manual, plan, resume]
...
    ValveStateEnum:
      type: string
      enum: [on, off, unknown]

    # ── Valve ─────────────────────────────────────────────────────────────────
...
    ZoneSummary:
      type: object
      required: [id, label, valve_state]
      properties:
        id:          { type: string }
        label:       { type: string }
        valve_state: { $ref: '#/components/schemas/ValveStateEnum' }
    CurrentRun:
      type: object
      required: [zone_id, zone_label, started_at, elapsed_min, remaining_min, total_min, triggered_by]
      properties:
        zone_id:       { type: string }
        zone_label:    { type: string }
        started_at:    { type: string, format: date-time }
        elapsed_min:   { type: number }
        remaining_min: { type: number }
        total_min:     { type: number }
        triggered_by:  { $ref: '#/components/schemas/TriggerSource' }
    ScheduleSummary:
      type: object
      required: [days, times, cron]
      properties:
        days:   { type: array, items: { type: integer }, description: "ISO weekdays 1=Mon..7=Sun; empty = every day" }
        times:  { type: array, items: { type: string }, description: "Clock times HH:MM" }
        cron:   { type: string, nullable: true, description: "Advanced cron override; set only when active" }
    SequenceState:
      type: object
      required: [id, label, status, enabled, paused, factor_pct, factor_notes, schedule, next_run_at, zones, basis_min_per_zone, current_run]
      properties:
        id:                 { type: string }
        label:              { type: string }
        status:             { $ref: '#/components/schemas/SequenceStatus' }
        enabled:            { type: boolean }
        paused:             { type: boolean }
        factor_pct:         { type: integer }
        factor_notes:       { $ref: '#/components/schemas/FactorNotes' }
        schedule:           { $ref: '#/components/schemas/ScheduleSummary' }
        next_run_at:        { type: string, format: date-time, nullable: true }
        zones:
          type: array
          items: { $ref: '#/components/schemas/ZoneSummary' }
        basis_min_per_zone: { type: integer }
        current_run:
          nullable: true
          oneOf:
            - { $ref: '#/components/schemas/CurrentRun' }
            - { type: 'null' }
...
    StartZoneRequest:
      type: object
      required: [duration_min]
      properties:
        duration_min:
          type: integer
          minimum: 1
          description: "Run duration in minutes for this single zone (factor not applied)."

    # ── Status ────────────────────────────────────────────────────────────────
...
    FactorNotes:
      type: object
      description: Structured (machine-readable) reasons behind a sequence's factor; the frontend localizes them.
      required: [season_off, wind_blocked, rain_factor_pct, temp_delta_pct]
      properties:
        season_off:      { type: boolean }
        wind_blocked:    { type: boolean }
        rain_factor_pct: { type: integer, nullable: true, description: "Rain multiplier in % (100 = neutral); set only when rain reduces watering." }
        temp_delta_pct:  { type: integer, nullable: true, description: "Signed temperature delta in %; set only when |delta| >= 5." }
...
    TempFactorSettings:
      type: object
      required: [basis_c, pct_per_c, min_pct, max_pct]
      properties:
        basis_c:    { type: number }
        pct_per_c:  { type: number }
        min_pct:    { type: integer }
        max_pct:    { type: integer }
    RainFactorSettings:
      type: object
      required: [mode, forecast_days, threshold_prob, reduce_above_mm, zero_above_mm, forecast_decay, water_balance_days, water_balance_decay, peak_tomorrow, confirm_with_rain_sensor]
      properties:
        mode:
          type: string
          enum: [forecast, water_balance]
        forecast_days:     { type: integer }
        threshold_prob:    { type: integer }
        reduce_above_mm:   { type: number }
        zero_above_mm:     { type: number }
        forecast_decay:    { type: number }
        water_balance_days: { type: integer, description: "Recent actual-rain history window for water_balance mode." }
        water_balance_decay: { type: number, description: "Daily carryover for actual rain in water_balance mode." }
        peak_tomorrow:     { type: boolean, description: "When true the rain factor also uses tomorrow's daily peak forecast, not just the latest reading. Today always uses the peak." }
        confirm_with_rain_sensor: { type: boolean, description: "When true today's forecast peak only applies up to the level that coincided with the binary rain sensor actually detecting rain. In water_balance mode, actual-rain deltas also only count while the rain sensor was on." }
    SequenceOverride:
      type: object
      required: [basis_min_per_zone, watchdog_min, paused]
      properties:
        basis_min_per_zone: { type: integer, nullable: true }
        watchdog_min:       { type: integer, nullable: true }
        paused:             { type: boolean }
    TempFactorSettingsUpdate:
      type: object
      properties:
        basis_c:    { type: number }
        pct_per_c:  { type: number }
        min_pct:    { type: integer }
        max_pct:    { type: integer }
    RainFactorSettingsUpdate:
      type: object
      properties:
        mode:
          type: string
          enum: [forecast, water_balance]
        forecast_days:     { type: integer }
        threshold_prob:    { type: integer }
        reduce_above_mm:   { type: number }
        zero_above_mm:     { type: number }
        forecast_decay:    { type: number }
        water_balance_days: { type: integer, description: "Recent actual-rain history window for water_balance mode." }
        water_balance_decay: { type: number, description: "Daily carryover for actual rain in water_balance mode." }
        peak_tomorrow:     { type: boolean, description: "When true the rain factor also uses tomorrow's daily peak forecast, not just the latest reading. Today always uses the peak." }
        confirm_with_rain_sensor: { type: boolean, description: "When true today's forecast peak only applies up to the level that coincided with the binary rain sensor actually detecting rain. In water_balance mode, actual-rain deltas also only count while the rain sensor was on." }
    SequenceOverrideUpdate:
      type: object
      properties:
        basis_min_per_zone: { type: integer, nullable: true }
        watchdog_min:       { type: integer, nullable: true }
        paused:             { type: boolean }
    FactorSettingsUpdate:
      type: object
      properties:
        temp: { $ref: '#/components/schemas/TempFactorSettingsUpdate' }
        rain: { $ref: '#/components/schemas/RainFactorSettingsUpdate' }
        manual_mode: { type: boolean, description: "When true, the manual adjustment overrides the automatic factor." }
        manual_pct:  { type: integer, nullable: true, description: "Manual adjustment percentage (clamped to the temperature factor's min/max)." }
    UpdateSettingsRequest:
      type: object
      properties:
        sequences:
          type: object
          additionalProperties: { $ref: '#/components/schemas/SequenceOverrideUpdate' }
        factors: { $ref: '#/components/schemas/FactorSettingsUpdate' }
        token_lifetime_days:  { type: integer, minimum: 1, maximum: 365 }
        auto_login_enabled:   { type: boolean }
    AppSettings:
      type: object
      required: [sequences, factors, token_lifetime_days, auto_login_enabled]
      properties:
        sequences:
          type: object
          additionalProperties: { $ref: '#/components/schemas/SequenceOverride' }
        factors:
          type: object
          required: [temp, rain, manual_mode, manual_pct, temp_overridden, rain_overridden]
          properties:
            temp: { $ref: '#/components/schemas/TempFactorSettings' }
            rain: { $ref: '#/components/schemas/RainFactorSettings' }
            manual_mode: { type: boolean, description: "When true, the manual adjustment overrides the automatic factor." }
            manual_pct:  { type: integer, nullable: true, description: "Manual adjustment percentage (clamped to the temperature factor's min/max)." }
            temp_overridden: { type: boolean, description: "True when any temperature factor value is overridden in the DB (differs from the configured base)." }
            rain_overridden: { type: boolean, description: "True when any rain factor value is overridden in the DB (differs from the configured base)." }
        token_lifetime_days:  { type: integer }
        auto_login_enabled:   { type: boolean }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /sequences/{sequence_id}:
    get:
      operationId: getSequence
      summary: Get a single sequence
      parameters:
        - name: sequence_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SequenceState' }
        '404':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /zones/{zone_id}/start:
    post:
      operationId: startZone
      summary: Start a single zone immediately (in isolation, factor not applied)
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartZoneRequest' }
      responses:
        '202': { description: Start accepted }
        '404':
          description: Zone not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Mutex conflict — a sequence or zone is already running
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: Zone has no switch entity or master switch is off
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /zones/{zone_id}/stop:
    post:
      operationId: stopZone
      summary: Stop a standalone single-zone run
      parameters:
        - name: zone_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202': { description: Stop accepted }
        '409':
          description: Zone is not running as a standalone run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── Valves ──────────────────────────────────────────────────────────────────
...
  /settings:
    get:
      operationId: getSettings
      summary: Effective settings (YAML defaults merged with DB overrides)
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
    patch:
      operationId: updateSettings
      summary: Update settings (partial — only provided fields are overwritten)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSettingsRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
  /settings/factors:
    delete:
      operationId: clearFactorOverrides
      summary: Reset factor overrides to the configured base values
      parameters:
        - name: group
          in: query
          required: false
          description: Limit the reset to one factor group; omit to reset both.
          schema: { type: string, enum: [temp, rain] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
...
  /settings/sequences/{sequence_id}:
    delete:
      operationId: clearSequenceOverride
      summary: Remove the override for a single sequence (restore its YAML defaults)
      parameters:
        - name: sequence_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AppSettings' }
        '404':
          description: Unknown sequence

  # ── User preferences ────────────────────────────────────────────────────────
...